feat(ghostty): add ctrl-backspace binding
It is possible to delete a word in terminal (and terminal apps such as
neovim) using Ctrl+W. However, this keybinding closes tabs in almost all
other apps. Getting used to it might enable mistakes to happen in those
apps. In the rest of the system, Ctrl+backspace deletes a word, so let's
bind it in Ghostty as well. This makes sure we're consistent with the
rest of the system.